31/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sd-31"
version = "0.1.0"
description = "Software Design誌 第31回 デモアプリケーション"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.52.1",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langchain-anthropic>=1.2.0",
    "langgraph>=1.0.4",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
]

[tool.uv]
dev-dependencies = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "pytest>=9.0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
